rename: 디렉토리 이름 변경(auth -> cookie)
